meta-openembedded: farsight: fix implicit pkgconfig dependency causing configure fail
In the config log we see:

  configure:12277: checking whether to build static libraries
  configure:12281: result: yes
  configure:12419: checking for x86_64-overc-linux-pkg-config
  configure:12452: result: no
  configure:12462: checking for pkg-config
  configure:12495: result: no
  configure:12801: Using Farsight2 source release as package name
  configure:12825: Using Unknown package origin as package origin
  configure:12879: Using /usr/libexec/gst-install-plugins-helper as plugin install helper
  configure:12953: checking for NICE
  configure:13022: error: Need libnice >= 0.0.9

While it libnice that fails configure, it is the absence of pkg-config
several lines above that causes the failure, as libnice is in the sysroot.

For most build hosts, you'll have to remove the default binary that
lives in /usr/bin/pkg-config in order to see this failure.

Cc: Martin Jansa <Martin.Jansa@gmail.com>
(From meta-openembedded rev: 0d77dcf5c81d511bcefad6670a1dd2726f1ff2da)

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>